src/main/java/ua/ithillel/junit/calculator/SimpleCalculator.java
package ua.ithillel.junit.calculator;

public class SimpleCalculator {
    public double add(double a, double b) {
        return a + b;
    }

    public double subtract(double a, double b) {
        return a - b;
    }

    public double multiply(double a, double b) {
        return a * b;
    }

    public double divide(double a, double b) throws CalculatorException {
        if (b == 0) {
            throw new CalculatorException("Division by zero is not allowed");
        }
        return a / b;
    }
}
